Extract tests for isError to own file
